# Updated inits of heracles_analysis table to use more performant insert