Using many single inserts is very slow on Redshift and PDW, so I've changed these two init scripts to use INSERT INTO SELECT syntax, which performs much better. Tested on Redshift, PDW, and Postgres.

**Previous version benchmarks:**
- Redshift: 406 seconds
- PDW: 30 seconds
- Postgres: 0.39 seconds

**New version benchmarks:**
- Redshift: 35 seconds (a pure CTAS runs better: 18.07 seconds; perhaps we can merge the DDLs with the init scripts in the future)
- PDW: 0.79 seconds
- Postgres: 0.03 seconds

The reason I'm concerned about this performance is because we will be pulling all of the results DDLs and scripts hosted in this repo when executing the [CdmAtlasCutover](https://github.com/OHDSI/CdmAtlasCutover) package, which we use to cutover new CDMs into Atlas.