@alondhe 2018-07-18
Ah okay, makes sense @chrisknoll. In that case, we can just use these INSERT INTO SELECTs.