@alondhe 2018-08-01
Hi @anthonysena -- I added those field names to the inserts. Do I need to close and open a new PR, or does this PR have the new commit?